+++ packages/core/src/components/label-group/label-group.template.ts
@@ -1,4 +1,4 @@
-// Keep in sync packages\core\src\components\label-group\label-group.template.ts (https://github.com/dso-toolkit/dso-toolkit/issues/1438)
+// Keep in sync packages\css\src\components\label-group\label-group.template.ts (https://github.com/dso-toolkit/dso-toolkit/issues/1438)
 import { LabelGroup } from '@dso-toolkit/sources';
 import { html } from 'lit-html';
 import { labelTemplate } from '../label/label.template';