+++ packages/core/src/components/list/list.template.ts
@@ -1,4 +1,4 @@
-// Keep in sync with packages\core\src\components\list\list.template.ts (https://github.com/dso-toolkit/dso-toolkit/issues/1438)
+// Keep in sync with packages\css\src\components\list\list.template.ts (https://github.com/dso-toolkit/dso-toolkit/issues/1438)
 import { List, Type } from '@dso-toolkit/sources';
 
 import { html, TemplateResult } from 'lit-html';